+++ src/lib/Components/Bidding/Components/Checkbox.tsx
@@ -1,11 +1,10 @@
 import React, { Component } from "react"
-import { TouchableWithoutFeedback, TouchableWithoutFeedbackProperties } from "react-native"
+import { StyleSheet, TouchableWithoutFeedback, TouchableWithoutFeedbackProperties } from "react-native"
 import styled from "styled-components/native"
 
 import { Flex, FlexProps } from "../Elements/Flex"
 import { theme } from "../Elements/Theme"
-import { Sans12 } from "../Elements/Typography"
-import { Fade } from "./Animation/Fade"
+import { CssTransition } from "./Animation/CssTransition"
 
 interface CheckboxState {
   checked: boolean
@@ -14,21 +13,39 @@ interface CheckboxState {
 interface CheckboxProps extends TouchableWithoutFeedbackProperties, FlexProps {
   checked?: boolean
   disabled?: boolean
-  error?: string
+  error?: boolean
 }
 
 export class Checkbox extends Component<CheckboxProps, CheckboxState> {
-  // The size of the checkbox
   private readonly checkboxSize = 20
 
-  // Default border color
-  private readonly defaultBorderColor = "black10"
+  private readonly duration = 250
 
-  // The background color for the checkbox square box when `props.disabled` evaluates to `true`
-  private readonly bgColorForDisabledState = "black5"
+  private readonly defaultCheckboxStyle = {
+    backgroundColor: theme.colors.white100,
+    borderColor: theme.colors.black10,
+  }
 
-  // The color for the checkbox square box and the message when `props.error` is present
-  private readonly borderColorForErrorState = "red100"
+  private readonly checkedCheckboxStyle = {
+    backgroundColor: theme.colors.black100,
+    borderColor: theme.colors.black100,
+  }
+
+  private readonly disabledCheckboxStyle = {
+    backgroundColor: theme.colors.black5,
+    borderColor: theme.colors.black10,
+  }
+
+  private readonly checkboxStyles = {
+    default: {
+      unchecked: this.defaultCheckboxStyle,
+      checked: this.checkedCheckboxStyle,
+    },
+    error: {
+      unchecked: Object.assign({}, this.defaultCheckboxStyle, { borderColor: theme.colors.red100 }),
+      checked: this.checkedCheckboxStyle,
+    },
+  }
 
   constructor(props) {
     super(props)
@@ -50,50 +67,65 @@ export class Checkbox extends Component<CheckboxProps, CheckboxState> {
 
   render() {
     const { error, disabled, children, ...props } = this.props
+    const { checked } = this.state
 
-    const borderColor = error ? this.borderColorForErrorState : this.defaultBorderColor
-    const bg = disabled ? this.bgColorForDisabledState : null
+    let checkboxStyle
+    if (disabled) {
+      checkboxStyle = this.disabledCheckboxStyle
+    } else {
+      checkboxStyle = this.checkboxStyles[error ? "error" : "default"][checked ? "checked" : "unchecked"]
+    }
 
     return (
       <TouchableWithoutFeedback onPress={disabled ? () => null : this.onPress}>
-        <Flex flexDirection="column" {...props}>
-          <Flex mb={2} alignItems="center" flexDirection="row">
-            <SquareBox borderColor={borderColor} bg={bg} width={this.checkboxSize} height={this.checkboxSize}>
-              <Fade show={this.state.checked} duration={250}>
-                {disabled ? <Disabled size={this.checkboxSize} /> : <Enabled size={this.checkboxSize} />}
-              </Fade>
-            </SquareBox>
+        <Flex alignItems="center" flexDirection="row" {...props}>
+          <CssTransition
+            style={[styles.container, checkboxStyle]}
+            animate={["backgroundColor", "borderColor"]}
+            duration={this.duration}
+          >
+            {checked && (disabled ? <DisabledMark size={this.checkboxSize} /> : <CheckMark size={this.checkboxSize} />)}
+          </CssTransition>
 
-            <Flex>{children}</Flex>
-          </Flex>
-          {error && <Sans12 color="red100">{error}</Sans12>}
+          <Flex>{children}</Flex>
         </Flex>
       </TouchableWithoutFeedback>
     )
   }
 }
 
-const SquareBox = props => <Flex justifyContent="center" alignItems="center" mr={3} border="2px solid" {...props} />
+// styled-component does not have support for Animated.View
+const styles = StyleSheet.create({
+  container: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    borderWidth: 2,
+    borderStyle: "solid",
+    marginRight: theme.space[3],
+    width: 20,
+    height: 20,
+  },
+})
 
-interface CheckmarkPropss {
+interface CheckMarkPropss {
   size: number
 }
 
 // This component represents the √ mark in CSS. We are not using styled-system since it's easier to specify raw CSS
 // properties with styled-component.
-const Enabled = styled.View.attrs<CheckmarkPropss>({})`
+export const CheckMark = styled.View.attrs<CheckMarkPropss>({})`
   transform: rotate(-45deg);
-  top: -25%;
-  left: -3%;
+  top: -12%;
   width: ${props => props.size * 0.625};
   height: ${props => props.size * 0.3125};
-  border-bottom-color: black;
+  border-bottom-color: white;
   border-bottom-width: 2px;
-  border-left-color: black;
+  border-left-color: white;
   border-left-width: 2px;
 `
 
-const Disabled = Enabled.extend`
+export const DisabledMark = CheckMark.extend`
   border-bottom-color: ${theme.colors.black30};
   border-left-color: ${theme.colors.black30};
 `