+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * Adds `<MarkdownRenderer>` component - sepans
 * Adds a `<Input>` component - yuki24
 * Adds a billing address screen - yuki24
+* Reflect style updates for the `<Checkbox>` component - yuki24
 
 ### 1.4.6